+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 **Added**
 
-- Added `Coordinator#getRolesByOrganization` for getting all roles belonging to an organization
+- Added Roles Module `coordinator.roles`
+- Added `Roles#getByOrganizationId` for getting all roles belonging to an organization
 
 ## [v0.0.49](http://github.com/ndustrialio/contxt-sdk-js/tree/v0.0.49) (2019-04-22)